+++ src/main/groovy/com/monits/gradle/sca/config/PmdConfigurator.groovy
@@ -22,6 +22,8 @@ import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.plugins.quality.Pmd
+import org.gradle.api.tasks.SourceSet
+import org.gradle.api.tasks.compile.JavaCompile
 import org.gradle.util.GUtil
 import org.gradle.util.GradleVersion
 
@@ -36,7 +38,14 @@ class PmdConfigurator implements AnalysisConfigurator, ClasspathAware {
     void applyConfig(final Project project, final StaticCodeAnalysisExtension extension) {
         setupPlugin(project, extension)
 
-        setupTasksPerSourceSet(project, extension, project.sourceSets)
+        setupTasksPerSourceSet(project, extension, project.sourceSets) { Pmd pmdTask, SourceSet sourceSet ->
+            boolean supportsClasspath = GRADLE_VERSION_PMD_CLASSPATH_SUPPORT <= GradleVersion.current()
+
+            if (supportsClasspath) {
+                // This is the default in Gradle 3.+, we backport it
+                pmdTask.classpath = sourceSet.output + sourceSet.compileClasspath
+            }
+        }
     }
 
     // DuplicateStringLiteral should be removed once we refactor this
@@ -45,6 +54,13 @@ class PmdConfigurator implements AnalysisConfigurator, ClasspathAware {
         setupPlugin(project, extension)
 
         setupTasksPerSourceSet(project, extension, project.android.sourceSets) { Pmd pmdTask, sourceSet ->
+            /*
+             * Android doesn't expose name of the task compiling the sourceset, and names vary
+             * widely from version to version of the plugin, plus needs to take flavors into account.
+             * This is inefficient, but safer and simpler.
+            */
+            dependsOn project.tasks.withType(JavaCompile)
+
             source sourceSet.java.srcDirs
             exclude '**/gen/**'